+++ src/mg/emberframework/util/validation/validator/RequiredValidator.java
@@ -1,16 +1,16 @@
-package mg.emberframework.util.validation;
+package mg.emberframework.util.validation.validator;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 
 import mg.emberframework.manager.exception.ModelValidationException;
 
-public class NumericValidator implements ValidatorInterface{
+public class RequiredValidator implements ValidatorInterface{
 
     @Override
     public void validate(String value, Annotation annotation, Field field) throws ModelValidationException {
-        if (!(value != null && value.matches("\\d+"))) {
-            throw new ModelValidationException(field.getName() + " should be numeric");
+        if (value == null || "".equals(value)) {
+            throw new ModelValidationException(field.getName() + " should is Required");
         }
     }
