+++ src/mg/emberframework/util/validation/validator/ValidatorInterface.java
@@ -1,4 +1,4 @@
-package mg.emberframework.util.validation;
+package mg.emberframework.util.validation.validator;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
